recruit.ai/requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
imap-tools
python-multipart
pypdf
python-docx
python-dotenv
mail-parser
mammoth
jinja2
pinecone
